requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
filetype==1.2.0
fonttools==4.58.1
idna==3.7
kiwisolver==1.4.8
matplotlib==3.8.2
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.50
six==1.17.0
tqdm==4.67.1
urllib3==2.4.0
